Extract Class: FrameBuffer_RenderTarget is put in it's own class.
