+++ src/RageUtil/CMakeLists.txt
@@ -50,6 +50,7 @@ list(APPEND SMDATA_RAGE_GRAPHICS_SRC
   "Graphics/Display/RageDisplay_Null.cpp"
   "Graphics/Display/OpenGL/RageDisplay_OGL.cpp"
   "Graphics/Display/OpenGL/RageDisplay_OGL_Helpers.cpp"
+  "Graphics/Display/OpenGL/FrameBufferRenderTargetGL.cpp"
   "Graphics/Display/OpenGL/ShaderGL.cpp"
   "Graphics/Display/OpenGL/GLHelper.cpp"
   "Graphics/RageModelGeometry.cpp"
@@ -74,6 +75,7 @@ list(APPEND SMDATA_RAGE_GRAPHICS_HPP
   "Graphics/Display/RageDisplay_Null.h"
   "Graphics/Display/OpenGL/RageDisplay_OGL.h"
   "Graphics/Display/OpenGL/RageDisplay_OGL_Helpers.h"
+  "Graphics/Display/OpenGL/FrameBufferRenderTargetGL.hpp"
   "Graphics/Display/OpenGL/ShaderGL.hpp"
   "Graphics/Display/OpenGL/GLHelper.hpp"
   "Graphics/RageModelGeometry.h"
