+++ src/RageUtil/Graphics/Display/RageDisplay.h
@@ -35,10 +35,10 @@ class RageCompiledGeometry
   public:
 	virtual ~RageCompiledGeometry();
 
-	void Set(const vector<msMesh>& vMeshes, bool bNeedsNormals);
+	void Set(const std::vector<msMesh>& vMeshes, bool bNeedsNormals);
 
-	virtual void Allocate(const vector<msMesh>& vMeshes) = 0; // allocate space
-	virtual void Change(const vector<msMesh>& vMeshes) = 0; // new data must be
+	virtual void Allocate(const std::vector<msMesh>& vMeshes) = 0; // allocate space
+	virtual void Change(const std::vector<msMesh>& vMeshes) = 0; // new data must be
 															// the same size as
 															// was passed to
 															// Set()
@@ -70,7 +70,7 @@ class RageCompiledGeometry
 		int iTriangleCount;
 		bool m_bNeedsTextureMatrixScale;
 	};
-	vector<MeshInfo> m_vMeshInfo;
+	std::vector<MeshInfo> m_vMeshInfo;
 	bool m_bNeedsNormals{};
 	bool m_bAnyNeedsTextureMatrixScale{};
 };
